better test readiness
