+++ core/node/state_keeper/src/lib.rs
@@ -25,3 +25,4 @@ pub(crate) mod tests;
 pub(crate) mod types;
 pub mod updates;
 pub(crate) mod utils;
+mod v26_utils;
