+++ core/node/state_keeper/src/mempool_actor.rs
@@ -11,23 +11,13 @@ use zksync_mempool::L2TxFilter;
 use zksync_multivm::{
     utils::derive_base_fee_and_gas_per_pubdata,
     vm_fast::interface::opcodes::Add,
-    vm_latest::utils::v26_upgrade::{is_unsafe_deposit_present, AsyncStorageKeyAccess},
 };
 use zksync_node_fee_model::BatchFeeModelInputProvider;
 use zksync_types::{
-    address_to_h256,
-    ethabi::{self, Param, ParamType, Token},
-    get_address_mapping_key, get_immutable_simulator_key, get_nonce_key, h256_to_address,
-    hasher::keccak,
-    tx::execute::Create2DeploymentParams,
-    utils::encode_ntv_asset_id,
-    vm::VmVersion,
-    web3::keccak256,
-    AccountTreeId, Address, Nonce, StorageKey, Transaction, TransactionTimeRangeConstraint, H256,
-    L2_ASSET_ROUTER_ADDRESS, L2_ASSET_ROUTER_LEGACY_SHARED_BRIDGE_IMMUTABLE_KEY,
-    L2_LEGACY_SHARED_BRIDGE_BEACON_PROXY_BYTECODE_KEY, L2_LEGACY_SHARED_BRIDGE_L1_ADDRESSES_KEY,
-    L2_NATIVE_TOKEN_VAULT_ADDRESS, L2_NATIVE_TOKEN_VAULT_ASSET_ID_MAPPING_INDEX,
+    address_to_h256, ethabi::{self, Param, ParamType, Token}, get_address_mapping_key, get_immutable_simulator_key, get_nonce_key, h256_to_address, h256_to_u256, hasher::keccak, tx::execute::Create2DeploymentParams, utils::encode_ntv_asset_id, vm::VmVersion, web3::keccak256, AccountTreeId, Address, Nonce, StorageKey, Transaction, TransactionTimeRangeConstraint, H256, L2_ASSET_ROUTER_ADDRESS, L2_ASSET_ROUTER_LEGACY_SHARED_BRIDGE_IMMUTABLE_KEY, L2_ASSET_ROUTER_LEGACY_SHARED_BRIDGE_L1_CHAIN_ID_KEY, L2_LEGACY_SHARED_BRIDGE_BEACON_PROXY_BYTECODE_KEY, L2_LEGACY_SHARED_BRIDGE_L1_ADDRESSES_KEY, L2_LEGACY_SHARED_BRIDGE_UPGRADEABLE_BEACON_ADDRESS_KEY, L2_NATIVE_TOKEN_VAULT_ADDRESS, L2_NATIVE_TOKEN_VAULT_ASSET_ID_MAPPING_INDEX, U256
 };
+
+use crate::v26_utils::is_unsafe_deposit_present;
 
 use super::{metrics::KEEPER_METRICS, types::MempoolGuard};
 
@@ -58,27 +48,6 @@ pub struct MempoolFetcher {
     #[cfg(test)]
     transaction_hashes_sender: mpsc::UnboundedSender<Vec<H256>>,
 }
-struct PostgresStorageKeyAccess<'a, 'b> {
-    storage: &'a mut Connection<'b, Core>,
-}
-
-impl<'a, 'b> PostgresStorageKeyAccess<'a, 'b> {
-    fn new(storage: &'a mut Connection<'b, Core>) -> Self {
-        Self { storage }
-    }
-}
-
-#[async_trait::async_trait]
-impl<'a, 'b> AsyncStorageKeyAccess for PostgresStorageKeyAccess<'a, 'b> {
-    async fn read_key(&mut self, key: &StorageKey) -> anyhow::Result<H256> {
-        self.storage
-            .storage_web3_dal()
-            .get_value(&key)
-            .await
-            .map_err(|e| e.into())
-    }
-}
-
 impl MempoolFetcher {
     pub fn new(
         mempool: MempoolGuard,
@@ -171,7 +140,7 @@ impl MempoolFetcher {
 
             let exclude_l1_txs = is_unsafe_deposit_present(
                 &transactions_with_constraints,
-                &mut PostgresStorageKeyAccess::new(&mut storage),
+                &mut storage,
             )
             .await?;
